Refactor json serialiation helpers into a new package for developer consumption
