+++ src/Fdc3.Json/MorganStanley.Fdc3.Json.csproj
@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<RootNamespace>MorganStanley.Fdc3</RootNamespace>
-		<AssemblyName>MorganStanley.Fdc3</AssemblyName>
+		<RootNamespace>MorganStanley.Fdc3.Json</RootNamespace>
+		<AssemblyName>MorganStanley.Fdc3.Json</AssemblyName>
 		<VersionPrefix>2.0.0</VersionPrefix>
         <VersionSuffix>alpha.1</VersionSuffix>
 		<Product>fdc3-dotnet</Product>
@@ -36,6 +36,10 @@
 	    <Pack>True</Pack>
 	    <PackagePath>\</PackagePath>
 	  </None>
+	</ItemGroup>
+
+	<ItemGroup>
+	  <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
 	</ItemGroup>
 
 </Project>
