+++ src/Fdc3.Json/Serialization/Fdc3JsonSerializerSettings.cs
@@ -1,0 +1,37 @@
+﻿/*
+ * Morgan Stanley makes this available to you under the Apache License,
+ * Version 2.0 (the "License"). You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * See the NOTICE file distributed with this work for additional information
+ * regarding copyright ownership. Unless required by applicable law or agreed
+ * to in writing, software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using Newtonsoft.Json.Serialization;
+
+namespace MorganStanley.Fdc3.Json.Serialization
+{
+    public class Fdc3JsonSerializerSettings : JsonSerializerSettings
+    {
+        public Fdc3JsonSerializerSettings()
+        {
+            this.TypeNameHandling = TypeNameHandling.None;
+            this.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
+            this.DefaultValueHandling = DefaultValueHandling.Populate;
+            this.NullValueHandling = NullValueHandling.Ignore;
+            this.ContractResolver = new DefaultContractResolver()
+            {
+                NamingStrategy = new Fdc3CamelCaseNamingStrategy()
+            };
+            this.Converters = new JsonConverter[] { new StringEnumConverter() };
+
+        }
+    }
+}
